localization-management-api/pyproject.toml
[project]
name = "localization-management-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Shishir Jakati",email = "shishir.jakati98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "supabase (>=2.15.2,<3.0.0)"
]

[tool.poetry]
packages = [{include = "localization_management_api", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
